+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This file contains a chronological list of all major changes made to this crate.
 - Add tokenizer/lexer for ~{..} style templates.
 - Partially implement template processor
 - This is what we in the industry call a mess
+- Added JS bundling processor using brk_rolldown (#10)
+
+### 🐛 Fixes
+
+- Compiler errors
+- Template tests
 
 ### Other Changes